Use a uint32 for block count to support large heaps
This increases overhead, but is a requirement to use megabytes worth of heap.

+++ src/umm_malloc.c
@@ -83,7 +83,7 @@ UMM_H_ATTPACKPRE typedef struct umm_block_t {
 struct umm_heap_config {
     umm_block *pheap;
     size_t heap_size;
-    uint16_t numblocks;
+    uint32_t numblocks;
 };
 
 struct umm_heap_config umm_heap_current;
